Revert "[Aio] Add AsyncIO support to grpcio-reflection"
