--- src/python/grpcio_tests/tests_aio/reflection/BUILD.bazel
@@ -1,30 +1,0 @@
-# Copyright 2020 gRPC authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-package(default_testonly = 1)
-
-py_test(
-    name = "reflection_servicer_test",
-    srcs = ["reflection_servicer_test.py"],
-    imports = ["../../"],
-    python_version = "PY3",
-    deps = [
-        "//src/proto/grpc/testing:empty_py_pb2",
-        "//src/proto/grpc/testing/proto2:empty2_extensions_proto",
-        "//src/proto/grpc/testing/proto2:empty2_proto",
-        "//src/python/grpcio/grpc:grpcio",
-        "//src/python/grpcio_reflection/grpc_reflection/v1alpha:grpc_reflection",
-        "//src/python/grpcio_tests/tests_aio/unit:_test_base",
-    ],
-)
