+++ src/python/grpcio_tests/tests_aio/tests.json
@@ -3,7 +3,6 @@
   "health_check.health_servicer_test.HealthServicerTest",
   "interop.local_interop_test.InsecureLocalInteropTest",
   "interop.local_interop_test.SecureLocalInteropTest",
-  "reflection.reflection_servicer_test.ReflectionServicerTest",
   "unit._metadata_test.TestTypeMetadata",
   "unit.abort_test.TestAbort",
   "unit.aio_rpc_error_test.TestAioRpcError",
